Refactored job notification
Notify method now exists for all proxy clients that inherit BaseProxyClient.
Divided and conquered JSON messages (message, response and notify job notification).
