+++ GetworkStratumProxy/Proxy/Client/BaseProxyClient.cs
@@ -1,6 +1,10 @@
-﻿using System;
+﻿using GetworkStratumProxy.Extension;
+using GetworkStratumProxy.Rpc;
+using System;
+using System.IO;
 using System.Net;
 using System.Net.Sockets;
+using System.Text.Json;
 
 namespace GetworkStratumProxy.Proxy.Client
 {
@@ -17,11 +21,21 @@ namespace GetworkStratumProxy.Proxy.Client
         public EndPoint Endpoint { get; private set; }
         public StratumState StratumState { get; protected set; }
 
+        protected StreamWriter BackgroundJobWriter { get; set; }
+
         public BaseProxyClient(TcpClient tcpClient)
         {
             TcpClient = tcpClient;
             Endpoint = tcpClient.Client.RemoteEndPoint;
             StratumState = StratumState.Unauthorised;
+        }
+
+        protected void Notify(JsonResponse notification)
+        {
+            var notificationString = JsonSerializer.Serialize(notification);
+            ConsoleHelper.Log(GetType().Name, $"(O) {notificationString} -> {Endpoint}", LogLevel.Debug);
+            BackgroundJobWriter.WriteLine(notificationString);
+            BackgroundJobWriter.Flush();
         }
 
         public abstract void Dispose();
