+++ GetworkStratumProxy/Rpc/JsonMessage.cs
@@ -1,0 +1,13 @@
+﻿using System.Text.Json.Serialization;
+
+namespace GetworkStratumProxy.Rpc
+{
+    public abstract class JsonMessage
+    {
+        [JsonPropertyName("id")]
+        public int? Id { get; set; }
+
+        [JsonPropertyName("jsonrpc")]
+        public string JsonRpc { get; set; }
+    }
+}
